% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compile_result}
\alias{compile_result}
\title{Compile Results from qpgee()}
\usage{
compile_result(qpgee_results, beta0, threshold = 10^-3)
}
\arguments{
\item{qpgee_results}{A (list of) fitted qpgee model.}

\item{beta0}{True beta used in true data generation process.}

\item{threshold}{Integer, the threshold to determine whether a esimated beta should
be consider as 0.}
}
\value{
a vector contains correct percentage, TP, FP, MSE and MAD and its standard
error if Monte Carlo simulations.
}
\description{
This function reports correct percentage, TP, FP, MSE and MAD from a (list of)
fitted qpgee model comparing to the true betas.
}
