% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgee.R
\name{.update_beta}
\alias{.update_beta}
\title{Update Beta Coefficients for One Iteration}
\usage{
.update_beta(
  x_active,
  y,
  mu,
  beta_current,
  R,
  nobs,
  tau,
  n_subjects,
  cluster_indices,
  lambda,
  penalty_factor_active,
  f0 = NULL
)
}
\description{
This is an internal helper function to compute the update step for beta.
}
\keyword{internal}
