% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeVerse.R
\name{geeVerse}
\alias{geeVerse}
\title{GeeVerse: Wrapper for Quantile Penalized Generalized Estimating Equations}
\usage{
geeVerse(x, y, tau = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{y}{A numeric vector of response variables.}

\item{tau}{A vector of quantiles to be estimated (default is c(0.25, 0.5, 0.75)).}

\item{...}{Additional arguments to be passed to qpgee function.}
}
\value{
A list containing the results for each tau value and a combined beta matrix.
}
\description{
This function is a wrapper for qpgee that allows running the model for multiple quantile levels (tau).
}
