% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEE.R
\name{PGEE}
\alias{PGEE}
\title{PGEE accelerated with RCpp}
\usage{
PGEE(
  formula,
  id,
  data,
  na.action = NULL,
  family = gaussian(link = "identity"),
  corstr = "independence",
  Mv = NULL,
  beta_int = NULL,
  R = NULL,
  scale.fix = TRUE,
  scale.value = 1,
  lambda,
  pindex = NULL,
  eps = 10^-6,
  maxiter = 30,
  tol = 10^-3,
  silent = TRUE
)
}
\arguments{
\item{formula}{A formula expression \code{response ~ predictors};}

\item{id}{A vector for identifying subjects/clusters.}

\item{data}{A data frame which stores the variables in \code{formula} with \code{id} variable.}
}
\value{
a PGEE object, which includes:
        fitted coefficients - the fitted single index coefficients with unit norm and first component being non negative
}
\description{
A function to fit penalized generalized estimating equation model.
This function was re-wrote partly with RCPP and RCPPEigen for better computation efficiency.
}
\examples{
#generate data
set.seed(2021)

n=n_sub=200
p=10
m=5
beta0=rep(1,m)
p0=length(beta0)
beta = c(beta0,rep(0,p-p0))
id = rep(1:n, each = m)
n_obs<-rep(m,n_sub);

e = NULL
for (i in 1:n_sub){
  sigmai=Siga_cov(0.5,"ar",n_obs[i])
  ei=mvtnorm::rmvnorm(1, mean=rep(0, n_obs[i]), sigma=sigmai)
  e=c(e,ei);
}

X=matrix(rnorm(n*m*p),n*m,p)
y=X\%*\%beta+(1+abs(X[,1]))*e

data = data.frame(X,y,id)

PGEE_fit = PGEE("y ~.-id-1",id = id, data = data,corstr = "exchangeable",lambda=0.01)
PGEE_fit$coefficients

}
