% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pp_scad_lin}
\alias{pp_scad_lin}
\title{SCAD Penalty Derivative (Linearized Approximation)}
\usage{
pp_scad_lin(beta, lambda, a = 3.7)
}
\arguments{
\item{beta}{A vector of coefficients.}

\item{lambda}{The main penalty tuning parameter.}

\item{a}{The second tuning parameter of the SCAD penalty (often 3.7).}
}
\value{
A vector of penalty derivative values.
}
\description{
This function calculates the derivative of the SCAD penalty, often used in
the local linear approximation algorithm.
}
\keyword{internal}
