% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.qpgee}
\alias{print.summary.qpgee}
\title{Print summary method for qpgee model objects}
\usage{
\method{print}{summary.qpgee}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A `qpgee` model object.}

\item{digits}{Default digits.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a summary of the qpgee model.
}
\description{
Print summary method for qpgee model objects
}
