% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuGene.R
\docType{data}
\name{simuGene}
\alias{simuGene}
\title{A Simulated Genetic Data from HapGen2}
\format{
A data frame with 1000 rows (subjects) and 500 columns (SNPs).
}
\usage{
simuGene
}
\description{
The `simuGene` dataset contains 500 SNPs simulated  data from
a commonly used tool for genetic data, HapGen2. We re-sampled existing genotype
data to create this simulated data. The genotype data we resample from is the
publicly available 1000 Genomes Project data. More specifically, we use resampled
from chromosome 14.
}
\examples{
data(simuGene)
head(simuGene)
}
\keyword{datasets}
